+++ pipenv/utils/pipfile.py
@@ -1,5 +1,14 @@
 import os
 
+from pipenv import exceptions, environments
+from pipenv.utils.requirements import import_requirements
+from pipenv.vendor import click
+
+from pipenv.patched.pip._vendor import rich
+
+console = rich.console.Console()
+err = rich.console.Console(stderr=True)
+
 
 def walk_up(bottom):
     """mimic os.walk, but walk 'up' instead of down the directory tree.
@@ -45,3 +54,72 @@ def find_pipfile(max_depth=3):
                 if os.path.isfile(p):
                     return p
     raise RuntimeError("No Pipfile found!")
+
+
+def ensure_pipfile(project, validate=True, skip_requirements=False, system=False):
+    """Creates a Pipfile for the project, if it doesn't exist."""
+
+    # Assert Pipfile exists.
+    python = (
+        project._which("python")
+        if not (project.s.USING_DEFAULT_PYTHON or system)
+        else None
+    )
+    if project.pipfile_is_empty:
+        # Show an error message and exit if system is passed and no pipfile exists
+        if system and not project.s.PIPENV_VIRTUALENV:
+            raise exceptions.PipenvOptionsError(
+                "--system",
+                "--system is intended to be used for pre-existing Pipfile "
+                "installation, not installation of specific packages. Aborting.",
+            )
+        # If there's a requirements file, but no Pipfile...
+        if project.requirements_exists and not skip_requirements:
+            requirements_dir_path = os.path.dirname(project.requirements_location)
+            click.echo(
+                "{0} found in {1} instead of {2}! Converting...".format(
+                    click.style("requirements.txt", bold=True),
+                    click.style(requirements_dir_path, fg="yellow", bold=True),
+                    click.style("Pipfile", bold=True),
+                )
+            )
+            # Create a Pipfile...
+            project.create_pipfile(python=python)
+            with console.status(
+                "Importing requirements...", spinner=project.s.PIPENV_SPINNER
+            ) as st:
+                # Import requirements.txt.
+                try:
+                    import_requirements(project)
+                except Exception:
+                    err.print(environments.PIPENV_SPINNER_FAIL_TEXT.format("Failed..."))
+                else:
+                    st.console.print(
+                        environments.PIPENV_SPINNER_OK_TEXT.format("Success!")
+                    )
+            # Warn the user of side-effects.
+            click.echo(
+                "{0}: Your {1} now contains pinned versions, if your {2} did. \n"
+                "We recommend updating your {1} to specify the {3} version, instead."
+                "".format(
+                    click.style("Warning", fg="red", bold=True),
+                    click.style("Pipfile", bold=True),
+                    click.style("requirements.txt", bold=True),
+                    click.style('"*"', bold=True),
+                )
+            )
+        else:
+            click.secho("Creating a Pipfile for this project...", bold=True, err=True)
+            # Create the pipfile if it doesn't exist.
+            project.create_pipfile(python=python)
+    # Validate the Pipfile's contents.
+    if validate and project.virtualenv_exists and not project.s.PIPENV_SKIP_VALIDATION:
+        # Ensure that Pipfile is using proper casing.
+        p = project.parsed_pipfile
+        changed = project.ensure_proper_casing()
+        # Write changes out to disk.
+        if changed:
+            click.echo(
+                click.style("Fixing package names in Pipfile...", bold=True), err=True
+            )
+            project.write_toml(p)
